+++ packages/react/ds/src/container/container.tsx
@@ -1,5 +1,3 @@
-import styles from './container.module.css';
-
 export function Container({ children }: { children: React.ReactNode }) {
-  return <div className={styles.container}>{children}</div>;
+  return <div className={`gi-mx-auto gi-container`}>{children}</div>;
 }