+++ packages/react/storybook/.storybook/main.ts
@@ -1,10 +1,7 @@
 import type { StorybookConfig } from '@storybook/react-vite';
 
 const config: StorybookConfig = {
-  stories: [
-    '../../ds/src/**/*.mdx',
-    '../../ds/src/**/*.stories.@(js|jsx|mjs|ts|tsx)',
-  ],
+  stories: ['../docs/*.mdx', '../../ds/src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
   addons: [
     '@storybook/addon-onboarding',
     '@storybook/addon-links',