+++ packages/react/ds/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [0.9.0-alpha.1](https://github.com/ogcio/govie-ds/compare/@govie-react/ds-0.9.0-alpha.0...@govie-react/ds-0.9.0-alpha.1) (2024-09-04)
+
+
+### Features
+
+* enable tailwind in react components ([#186](https://github.com/ogcio/govie-ds/issues/186)) ([f97f8b2](https://github.com/ogcio/govie-ds/commit/f97f8b2c5a370b580bd22246787cd6eea0d06997))
+
+
+
+# [0.9.0-alpha.0](https://github.com/ogcio/govie-ds/compare/@govie-react/ds-0.8.0...@govie-react/ds-0.9.0-alpha.0) (2024-09-04)
+
+
+### Features
+
+* paragraph component ([#195](https://github.com/ogcio/govie-ds/issues/195)) ([7794355](https://github.com/ogcio/govie-ds/commit/77943552d36c71de8cd237d64c52537347b0497d))
+
+
+
 # [0.8.0](https://github.com/ogcio/govie-ds/compare/@govie-react/ds-0.7.7...@govie-react/ds-0.8.0) (2024-08-06)
 
 ### Dependency Updates